Strip unused `MemoryReporter` out of Runtime Host API (#34363)
### Ticket
Closes #34135

### Problem description
Runtime is reducing it's host API surface area to prepare for an
upcoming refactoring.
`MemoryReporter` is identified to be an internal construct which is in
conflict of the public API folder it sits in.

### What's changed
Moved the `MemoryReporter` class from Runtime Host API into
`tt_metal/detail/reports/memory_reporter.hpp` where it's implementation
was defined.

### Checklist
- [x] [All post
commit](https://github.com/tenstorrent/tt-metal/actions/runs/20183647541)
- [x] No external dependency affected
- [x] No mlir dependency affected
