+++ tt_metal/api/tt-metalium/memory_reporter.hpp
@@ -1,22 +1,18 @@
-// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
 
 #pragma once
 
-#include <atomic>
 #include <cstddef>
 #include <cstdint>
-#include <filesystem>
-#include <fstream>
 #include <string>
 #include <unordered_map>
 #include <vector>
 
-#include <tt-metalium/allocator.hpp>
-
 namespace tt::tt_metal {
 enum class BufferType;
+class IDevice;
 }  // namespace tt::tt_metal
 
 namespace tt::tt_metal {
@@ -28,8 +24,6 @@ address: bytes
 size: bytes
 */
 using MemoryBlockTable = std::vector<std::unordered_map<std::string, std::string>>;
-class IDevice;
-class Program;
 
 namespace detail {
 struct MemoryView;
@@ -101,33 +95,6 @@ struct MemoryView {
     size_t total_bytes_free_per_bank = 0;
     size_t largest_contiguous_bytes_free_per_bank = 0;
     MemoryBlockTable block_table;
-};
-
-class MemoryReporter {
-public:
-    MemoryReporter& operator=(const MemoryReporter&) = delete;
-    MemoryReporter& operator=(MemoryReporter&& other) noexcept = delete;
-    MemoryReporter(const MemoryReporter&) = delete;
-    MemoryReporter(MemoryReporter&& other) noexcept = delete;
-
-    void flush_program_memory_usage(uint64_t program_id, const IDevice* device);
-
-    void dump_memory_usage_state(const IDevice* device, const std::string& prefix = "") const;
-
-    MemoryView get_memory_view(const IDevice* device, const BufferType& buffer_type) const;
-
-    static void toggle(bool state);
-    static MemoryReporter& inst();
-    static bool enabled();
-
-private:
-    MemoryReporter() = default;
-    ~MemoryReporter();
-    void init_reports();
-    static std::atomic<bool> is_enabled_;
-    std::ofstream program_l1_usage_summary_report_;
-    std::ofstream program_memory_usage_summary_report_;
-    std::ofstream program_detailed_memory_usage_report_;
 };
 
 }  // namespace detail
