+++ tt_metal/detail/reports/memory_reporter.cpp
@@ -1,10 +1,9 @@
-// SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
+// SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <allocator.hpp>
-#include <device.hpp>
-#include <memory_reporter.hpp>
+#include "tt_metal/detail/reports/memory_reporter.hpp"
+
 #include <cstdint>
 #include <filesystem>
 #include <map>
@@ -12,7 +11,9 @@
 #include <utility>
 #include <enchantum/enchantum.hpp>
 
-#include "buffer_types.hpp"
+#include <tt-metalium/allocator.hpp>
+#include <tt-metalium/device.hpp>
+#include <tt-metalium/buffer_types.hpp>
 #include "tt_metal/detail/reports/report_utils.hpp"
 #include "tt_metal/impl/allocator/allocator.hpp"
